Title: 앞에서부터 삭제하기 2, Time: 398ms, Memory: 22MB, Status: Passed - CodeTree

+++ 240407/앞에서부터 삭제하기 2/delete-it-from-the-beginning-2.java
@@ -30,7 +30,7 @@ public class Main {
 			int tmp = pq.poll();
 			sum -= tmp;
 
-			average = Math.max(average, sum / (N - k - 1));
+			average = Math.max(average, (double) sum / (N - k - 1));
 			pq.offer(tmp);
 			pq.offer(arr[k - 1]);
 			sum += tmp + arr[k - 1];
